% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_data}
\alias{covid_data}
\title{Covid cases and deaths for the Greater Region in 2020}
\format{
A data frame with 11277 rows and 6 variables:
\describe{
\item{day}{The day of the measurement, in the YYYY-MM-DD format}
\item{country}{Luxembourg, France, Belgium or Germany}
\item{region}{First administrative level. Lorraine for France, Wallonie for Belgium, Rheinland-Pfalz for Germany and Luxembourg for Luxembourg}
\item{sub_region}{Second administrative level: departments for France, provinces for Belgium, kreise for Germany and Luxembourg for Luxembourg}
\item{cases}{Positive cases}
\item{deaths}{Deaths}
...
}
}
\usage{
covid_data
}
\description{
A dataset with the daily cases for the lowest available administrative regions of the Greater Region.
Deaths are missing for France however. The \code{population} dataset can be used to obtain cases or
deaths per 100'000 inhabitants.
}
\keyword{datasets}
