% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_de_data}
\alias{get_de_data}
\title{Download data for the German border regions of Luxembourg}
\usage{
get_de_data(
  url = "https://opendata.arcgis.com/datasets/dd4580c810204019a7b8eb3e0b329dd6_0.csv",
  daily = TRUE
)
}
\arguments{
\item{url}{The url to the data. By default, points to the latest known url.}

\item{daily}{If TRUE, get daily cases, if FALSE, weekly cases.}
}
\value{
A data frame with the latest positive cases and deaths.
}
\description{
Download data for the German border regions of Luxembourg
}
\examples{
\dontrun{
get_de_data()
}
}
