% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_fr_data}
\alias{get_fr_data}
\title{Download data for the French border regions of Luxembourg}
\usage{
get_fr_data(
 
    url_alt = "https://raw.githubusercontent.com/opencovid19-fr/data/master/dist/chiffres-cles.csv",
 
    url_old = "https://www.data.gouv.fr/fr/datasets/r/b4ea7b4b-b7d1-4885-a099-71852291ff20",
 
    url_new = "https://www.data.gouv.fr/en/datasets/r/406c6a23-e283-4300-9484-54e78c8ae675",
  daily = TRUE
)
}
\arguments{
\item{url_alt}{The url to the data from the first wave (Spring 2020), but an alternative, more complete source. This
data is based on newspapers articles and daily reports from the French ARS. For the month of March, this data set
reflects the reality better. By default, points to the latest known url (source: \url{https://github.com/opencovid19-fr/data}).}

\item{url_old}{The url to the data from the first wave (Spring 2020). By default, points to the latest known url.}

\item{url_new}{The url to the data from after the first wave (Spring 2020). By default, points to the latest known url.}

\item{daily}{If TRUE, get daily cases, if FALSE, weekly cases.}
}
\value{
A data frame with the latest positive cases. Deaths are unavailable, so that column is empty.
}
\description{
Download data for the French border regions of Luxembourg
}
\examples{
\dontrun{
get_fr_data()
}
}
