% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_greater_region_data}
\alias{get_greater_region_data}
\title{Download data for the entirety of the Greater Region}
\usage{
get_greater_region_data(daily = TRUE)
}
\arguments{
\item{daily}{If TRUE, get daily cases, if FALSE, weekly cases.}
}
\value{
A data frame the latest positive cases data for the whole Greater Region
}
\description{
Download data for the entirety of the Greater Region
}
\examples{
\dontrun{
get_fr_data()
}
}
