% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_lu_data}
\alias{get_lu_data}
\title{Download data for Luxembourg.}
\usage{
get_lu_data(
  url = "https://data.public.lu/en/datasets/r/767f8091-0591-4b04-9a6f-a9d60cd57159",
  daily = TRUE
)
}
\arguments{
\item{url}{The url to the data. By default, points to the latest known url.}

\item{daily}{If TRUE, get daily cases, if FALSE, weekly cases.}
}
\value{
A data frame the latest positive cases data.
}
\description{
Download data for Luxembourg.
}
\examples{
\dontrun{
get_fr_data()
}
}
