% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grande_region_map}
\alias{grande_region_map}
\title{Map of the Greater Region}
\format{
An object of class \code{geo_list} of length 3.
}
\usage{
grande_region_map
}
\description{
Map of the Greater Region, loads a SpatialPolygonsDataFrame object.
}
\keyword{datasets}
