% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population of the region of the Greater Region}
\format{
A data frame with 46 rows and 6 variables:
\describe{
\item{country}{Luxembourg, France, Belgium or Germany}
\item{region}{First administrative level. Lorraine for France, Wallonie for Belgium, Rheinland-Pfalz and Saarland for Germany and Luxembourg for Luxembourg}
\item{sub_region}{Second administrative level: departments for France, provinces for Belgium, kreise for Germany and Luxembourg for Luxembourg}
\item{year}{The year the population was counted; (1st of January) 2020 is the only value}
\item{population}{Population of the sub_region}
\item{source}{Where these values can be downloaded}
...
}
}
\usage{
population
}
\description{
A dataset with the total population of the different regions (provinces for Wallonia,
Lankreise and Stadtkreise for the Rhineland-Palatinate, départements for France and
the country of Luxembourg) composing the Greater Region.
}
\keyword{datasets}
