% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOweek2date.R
\name{thursday0}
\alias{thursday0}
\title{Date of the nearest Thursday of a given date}
\usage{
thursday0(date)
}
\arguments{
\item{date}{Vector which can be coerced to class \code{Date}}
}
\value{
A vector of dates of the nearest Thursdays
}
\description{
This internal function returns the date of the Thursday of the week in which the given date is located.
}
\details{
The week starts on Monday and ends on Sunday.
}
\keyword{internal}
