% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOweek2date.R
\name{year0}
\alias{year0}
\title{Calendar year of a given date}
\usage{
year0(date)
}
\arguments{
\item{date}{Vector which can be coerced to class \code{Date}}
}
\value{
An integer vector of years
}
\description{
This internal function returns the year with century as integer.
}
\keyword{internal}
