% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_be_data}
\alias{get_be_data}
\title{Download data for the Belgian border regions of Luxembourg}
\usage{
get_be_data(
  url_cases = "https://epistat.sciensano.be/Data/COVID19BE_CASES_AGESEX.csv",
  url_deaths = "https://epistat.sciensano.be/Data/COVID19BE_MORT.csv",
  daily = TRUE
)
}
\arguments{
\item{url_cases}{The url to the cases data. By default, points to the latest known url.}

\item{url_deaths}{The url to the deaths data. By default, points to the latest known url.}

\item{daily}{If TRUE, get daily cases, if FALSE, weekly cases.}
}
\value{
A data frame the latest positive cases data.
}
\description{
Download data for the Belgian border regions of Luxembourg
}
\examples{
\dontrun{
get_be_data()
}
}
