% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_lu_data}
\alias{get_lu_data}
\title{Download data for Luxembourg.}
\usage{
get_lu_data(
  url = "https://data.public.lu/en/datasets/r/0c6efef2-7776-4cc5-9af9-51bc372c59b0",
  daily = TRUE
)
}
\arguments{
\item{url}{The url to the data. By default, points to the latest known url.}

\item{daily}{If TRUE, get daily cases and deaths, if FALSE, weekly cases.}
}
\value{
A data frame with the latest positive cases and deaths.
}
\description{
Download data for Luxembourg.
}
\examples{
\dontrun{
get_fr_data()
}
}
